Ver0.2/requirements.txt
# LangGraph and LangChain Core
langgraph>=0.2.20
langgraph-sdk>=0.1.0
langgraph-cli[inmem]>=0.3.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Data Processing and Analysis
pandas>=2.0.0
duckdb>=0.9.0
numpy>=1.24.0

# Web Framework for API (if needed)
fastapi>=0.104.0
uvicorn>=0.24.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: Advanced text analysis
textblob>=0.17.1
scikit-learn>=1.3.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0 

# Supabase
supabase>=1.0.0

#pyjwt
pyjwt>=2.0.0

# Production dependencies for Render deployment
psycopg2-binary>=2.9.0
redis>=5.0.0